Rearrange tests into integ/unit
