+++ .github/workflows/build_and_test.yml
@@ -47,14 +47,18 @@ jobs:
       - name: "Install ci dependencies"
         run: pip install -r requirements.txt --use-pep517
 
-      - name: Run tests
+      - name: Run unit tests
         run: pytest -v --cov=fauna --cov-context=test tests/unit
+
+      - name: Run integration tests
+        run: pytest -v --cov=fauna --cov-context=test tests/integration
         # To get more insight into tests which are only flaky when run in github actions -- use commands like below
         # run: env HTTPX_LOG_LEVEL=trace pytest --capture=no -v --cov=fauna --cov-context=test -k test_stream_max_open_streams
         env:
           FAUNA_ENDPOINT: "http://${{ matrix.fauna-docker-service.host }}:${{ matrix.fauna-docker-service.port }}"
           FAUNA_ROOT_KEY: secret
           USE_GITHUB_ACTION_OVERRIDES: 1
+
       - name: Generate coverage html report with dynamic contexts
         run: coverage html --show-contexts
