+++ tests/unit/test_client.py
@@ -1,9 +1,14 @@
 from datetime import timedelta
+from typing import Mapping
 
 import httpx
+import pytest
+import pytest_subtests
+from pytest_httpx import HTTPXMock
 
 import fauna
-from fauna import Client, HTTPXClient
+from fauna import Client, HTTPXClient, Header, fql
+from fauna.client import QueryOptions
 
 
 def test_client_defaults(monkeypatch):
@@ -86,3 +91,108 @@ def test_get_query_timeout():
 
     c = Client(query_timeout=timedelta(minutes=1))
     assert c.get_query_timeout() == timedelta(minutes=1)
+
+
+def test_query_options_set(
+    httpx_mock: HTTPXMock,
+    linearized: bool,
+    query_timeout_ms: int,
+    traceparent: str,
+    tags: Mapping[str, str],
+    max_contention_retries: int,
+):
+
+    def validate_headers(request: httpx.Request):
+        """
+        Validate each of the associated Headers are set on the request
+        """
+        assert request.headers[Header.Linearized] == str(linearized).lower()
+        # being explicit to not figure out encoding a Mapping in the test
+        assert request.headers[Header.Tags] == \
+            "project=teapot&hello=world&testing=foobar"
+        assert request.headers[Header.TimeoutMs] == f"{query_timeout_ms}"
+        assert request.headers[Header.Traceparent] == traceparent
+        assert request.headers[Header.MaxContentionRetries] \
+            == f"{max_contention_retries}"
+
+        return httpx.Response(
+            status_code=200,
+            json={"data": "mocked"},
+        )
+
+    httpx_mock.add_callback(validate_headers)
+
+    with httpx.Client() as mockClient:
+        c = Client(
+            http_client=HTTPXClient(mockClient),
+            tags={"project": "teapot"},
+        )
+
+        res = c.query(
+            fql("not used, just sending to a mock client"),
+            opts=QueryOptions(
+                query_tags=tags,
+                linearized=linearized,
+                query_timeout_ms=query_timeout_ms,
+                traceparent=traceparent,
+                max_contention_retries=max_contention_retries,
+            ),
+        )
+
+        assert res.status_code == 200
+
+
+def test_query_tags(
+    subtests: pytest_subtests.SubTests,
+    httpx_mock: HTTPXMock,
+):
+    expected = None
+
+    def validate_tags(request: httpx.Request):
+        if expected is not None:
+            assert request.headers[Header.Tags] == expected
+        else:
+            with pytest.raises(KeyError):
+                assert request.headers[Header.Tags] == ""
+
+        return httpx.Response(
+            status_code=200,
+            json={"data": "mocked"},
+        )
+
+    httpx_mock.add_callback(validate_tags)
+
+    with httpx.Client() as mockClient:
+        c = Client(
+            http_client=HTTPXClient(mockClient),
+            tags=None,
+        )
+
+        with subtests.test("should not be set"):
+            expected = None
+
+            c.tags.clear()
+            c.query(fql("not used, just sending to a mock client"))
+        with subtests.test("should be set on client"):
+            expected = "project=teapot"
+
+            c.tags.clear()
+            c.tags.update({"project": "teapot"})
+            c.query(fql("not used, just sending to a mock client"))
+        with subtests.test("should be set on query"):
+            expected = "silly=pants"
+
+            c.tags.clear()
+            c.query(
+                fql("not used, just sending to a mock client"),
+                QueryOptions(query_tags={"silly": "pants"}),
+            )
+        with subtests.test("should avoid conflicts"):
+            expected = "project=kettle"
+
+            c.tags.clear()
+            c.tags.update({"project": "teapot"})
+            c.query(
+                fql("not used, just sending to a mock client"),
+                QueryOptions(query_tags={"project": "kettle"}),
+            )
